move Errors.tsx to /components
