+++ src/main/typescript/components/form/DepositForm.tsx
@@ -28,7 +28,7 @@ import { ReduxAction } from "../../lib/redux"
 import { fetchMetadata, saveDraft, submitDeposit } from "../../actions/depositFormActions"
 import { AppState } from "../../model/AppState"
 import { DepositFormState } from "../../model/DepositForm"
-import { Alert, ReloadAlert } from "../../Errors"
+import { Alert, ReloadAlert } from "../Errors"
 import DepositLicenseForm from "./parts/DepositLicenseForm"
 import PrivacySensitiveDataForm from "./parts/PrivacySensitiveDataForm"
 import MessageForDataManagerForm from "./parts/MessageForDataManagerForm"
